+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+### Version 5.8.0 (18th August 2026)
+#### Added
+- Added `GetThirdPartySharingSettingsWithTimeout` to the `Adjust` API and `ThirdPartySharingSettingsChangedDelegate` to `AdjustConfig` for reading current third-party-sharing settings from the Adjust backend.
+- Added `IsDeviceIdsReadingEnabled` member to `AdjustConfig` as an umbrella member alternative to setting individual device ID disabling members.
+- Added support for Samsung MAPS (Mobile Application Preload System).
+- Added support for sending parameters that identify preinstalled Android system apps.
+
+#### Changed
+- Updated the Adjust Signature library version to 5.0.0.
+
+#### Native SDKs
+- [iOS@v5.8.0][ios_sdk_v5.8.0]
+- [Android@v5.8.0][android_sdk_v5.8.0]
+
+---
+
 ### Version 5.6.0 (20th April 2026)
 #### Added
 - Added support for remote trigger callbacks. You can now receive remote trigger updates by assigning the `RemoteTriggerDelegate` member of your `AdjustConfig` instance. Remote trigger payload is exposed as a JSON string.
@@ -1546,6 +1562,7 @@ Kudos to [Ivan](https://github.com/MatkovIvan) and [Evgeny](https://github.com/e
 [ios_sdk_v5.5.0]: https://github.com/adjust/ios_sdk/tree/v5.5.0
 [ios_sdk_v5.5.2]: https://github.com/adjust/ios_sdk/tree/v5.5.2
 [ios_sdk_v5.6.2]: https://github.com/adjust/ios_sdk/tree/v5.6.2
+[ios_sdk_v5.8.0]: https://github.com/adjust/ios_sdk/tree/v5.8.0
 
 [android_sdk_v3.5.0]: https://github.com/adjust/android_sdk/tree/v3.5.0
 [android_sdk_v4.1.0]: https://github.com/adjust/android_sdk/tree/v4.1.0
@@ -1611,6 +1628,7 @@ Kudos to [Ivan](https://github.com/MatkovIvan) and [Evgeny](https://github.com/e
 [android_sdk_v5.4.6]: https://github.com/adjust/android_sdk/tree/v5.4.6
 [android_sdk_v5.5.0]: https://github.com/adjust/android_sdk/tree/v5.5.0
 [android_sdk_v5.6.1]: https://github.com/adjust/android_sdk/tree/v5.6.1
+[android_sdk_v5.8.0]: https://github.com/adjust/android_sdk/tree/v5.8.0
 
 [windows_sdk_v4.12.0]: https://github.com/adjust/windows_sdk/tree/v4.12.0
 [windows_sdk_v4.13.0]: https://github.com/adjust/windows_sdk/tree/v4.13.0